# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Website Subscription Portal'
  description: 'Website Subscription Portal Apis'
  contact:
    name: 'Suraj Sharma'
    email: sharmasuraj41@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
  /api/create-post:
    post:
      tags:
        - Websites
      summary: 'Create new post for website'
      description: 'Create new post for website'
      operationId: users.createPost
      parameters:
        -
          $ref: '#/components/parameters/X-localization'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - website_id
                - title
                - description
              properties:
                website_id:
                  description: 'Website Id'
                  type: integer
                title:
                  type: string
                description:
                  type: string
              type: object
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiModel'
  /api/subscribe-website:
    post:
      tags:
        - Websites
      summary: 'Subscribe website by user'
      description: 'Subscribe website by user'
      operationId: users.subscribeWebsite
      parameters:
        -
          $ref: '#/components/parameters/X-localization'
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - website_id
                - user_id
              properties:
                website_id:
                  description: 'Website Id'
                  type: integer
                user_id:
                  description: 'User Id'
                  type: integer
              type: object
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiModel'
components:
  schemas:
    ApiModel:
      required:
        - statusCode
        - msg
        - data
      properties:
        statusCode:
          type: integer
          format: int32
        msg:
          type: string
        data:
          type: object
      type: object
      example:
        statusCode: 200
        message: successful
        data: null
    pageRequestSchema:
      properties:
        page:
          description: '(Optional) Pagination page number. Default is 1'
          type: integer
        limit:
          description: '(Optional) Pagination page limit. Default is 8'
          type: integer
      type: object
  parameters:
    X-localization:
      name: X-localization
      in: header
      description: 'Language (e.g. `en`, `ar`) (default will be: `en`)'
      schema:
        type: string
        enum:
          - en
          - ar
    page:
      name: page
      in: query
      description: '(Optional) Pagination page number. Default is 1'
      schema:
        type: integer
    limit:
      name: limit
      in: query
      description: '(Optional) Pagination page limit. Default is 8'
      schema:
        type: integer
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
      bearerFormat: JWT